Revert "feat sidebar (#14)"
This reverts commit 7c64d67ee60fbf8e3cbdf29843a18b7b9742771c.
