+++ api/index.ts
@@ -1,0 +1,3 @@
+export default function test() {
+  return;
+}
