+++ hooks/useTodoListContainer.ts
@@ -1,12 +1,28 @@
 import { todoKey } from 'constants/keyValue';
-import { useEffect } from 'react';
-
+import { useEffect, useState, useCallback, useRef } from 'react';
 import useTodoListStore from 'store/todoStore';
 import { getLocalStorage } from 'utils/localStorage';
 
 export default function useTodoListContainer() {
   const { todoList, setTodoList } = useTodoListStore();
-
+  const [date, setDate] = useState<Date>(new Date());
+  const [calendarVisible, setCalendarVisible] = useState<boolean>(false);
+  const calendarDivRef = useRef<HTMLDivElement>(null);
+  //수정중
+  const handleClickOutside = useCallback((e: MouseEvent) => {
+    const target = e.target as Node;
+    if (calendarDivRef.current && !calendarDivRef.current.contains(target)) {
+      setCalendarVisible(false);
+    }
+  }, []);
+  const handleOpenCalendar = useCallback((e: React.MouseEvent) => {
+    setCalendarVisible(true);
+  }, []);
+  const handleCloseCalendar = useCallback((e: React.MouseEvent) => {
+    const target = e.target as Node;
+    if (calendarDivRef.current && !calendarDivRef.current.contains(target))
+      setCalendarVisible(false);
+  }, []);
 
   useEffect(() => {
     //로컬 스토리지에서 todo
@@ -14,11 +30,17 @@ export default function useTodoListContainer() {
     if (todos) {
       setTodoList(JSON.parse(todos));
     }
-  }, [setTodoList]);
+  }, [handleClickOutside, setTodoList]);
 
   return {
+    calendarDivRef,
     todoList,
     setTodoList,
-
+    date,
+    setDate,
+    calendarVisible,
+    setCalendarVisible,
+    handleOpenCalendar,
+    handleCloseCalendar,
   };
 }
