+++ tailwind.config.js
@@ -85,31 +85,6 @@ module.exports = {
           from: { opacity: '1' },
           to: { opacity: '0' },
         },
-        calendarAppear: {
-          '0%': {
-            transformOrigin: '50% 0',
-            transform: 'scale(0.01, 0)',
-            opacity: '0'
-          },
-          '50%': {
-            transformOrigin: '50% 0',
-            transform: 'scale(0.01, 1)',
-            opacity: '1'
-          },
-          '100%': {
-            transform: 'scale(1, 1)',
-          },
-        },
-        calendarDisappear: {
-          '50%': {
-            transformOrigin: '50% 0',
-            transform: 'scale(0.01, 1)',
-            opacity: '1'
-          },
-          '100%': {
-            transformOrigin: '50% 0',
-            transform: 'scale(0.01, 0)',
-          },
         sliderUp: {
           from: { transform: 'translateY(100%)', opacity: '0' },
           to: { transform: 'translateY(0)', opacity: '1' },
@@ -122,7 +97,6 @@ module.exports = {
           '0%': { transform: 'translateY(100%)', opacity: '0' },
           '50%': { transform: 'translateY(0)', opacity: '1' },
           '100%': { transform: 'translateY(100%)', opacity: '0' },
-
         },
 
       },
@@ -139,8 +113,6 @@ module.exports = {
         profile: 'opacityDown 2s ease-in-out forwards',
         profileSecond: 'opacityDown 2s 2s ease-in-out forwards',
         profileAtter: 'opacityUp 3s 3s ease-in-out forwards',
-        calendarAppear:'calendarAppear 700ms ease-in-out both',
-        calendarDisappear:'calendarDisappear 700ms ease-in-out both'
         sliderUp: 'sliderUp 500ms ease-in-out forwards',
         sliderDown: 'sliderDown 500ms ease-in-out forwards',
         sliderMove: 'sliderMove 2s ease-in-out forwards',
